Move common option and flag logic into a mixin
Prefactor step before adding repeated option logic into Command and SubCommand builders.
